Move communication with native host to background
Before, sending and receiving the messages to and
from the native host is handled in popup. However
if the popup is closed before the response is
delivered, the callback is already destroyed.

Moving the communication code to the background
script solves this as it is always active. The
popup simply messages the background script when
the user clicks on 'fill'.